move _decode_basis_label() from basecircuit.py to quantum.py
